swipe controls V1
Amount ball is moved is based on distance of swipe. The further the swipe, the more force will be applied to the ball.

The swipe can take place anywhere on the screen (i.e., you don't have to swipe on the ball).
It is also possible to start to swipe in a direction, then keep your finger stationary, and the ball will gradually accelerate in that direction.