Movement code refactoring and marble clamped y-position.
Removed old and unused fields in BallController, moved some code from BallController to IBallMovementBehavior and BallSwipeMovement.

Marble's y-position is clamped 5 units above it's position when it is grounded. Once it leaves the ground, that clamped position stops updating, so when the ball reaches that clamped height it gets stopped.
IMPORTANT: The ball currently hangs in the air if it has a high y-velocity when it gets stopped. I'm lerping the y-velocity to 0 over .3 seconds to make it less jarring. This shouldn't be an issue since the ball should never have that high of a y-velocity during normal gameplay, but if we ever want to address how the ball "hangs" at the ceiling, refer to the script IBallMovementBehaviour.cs