Physic material changes, pyramid is now bouncy, changed pyramid size
Was struggling a lot earlier b/c the ball wasn't bouncing at all off the pyramid! Turns out the BounceThreshold in the Project Settings was set to 50, so any relative velocity below 50 wasn't being accounted for. Fixed this by changing it to 0.

I made the pyramid scale .9 on X, Y, and Z axes cause at 1 it was too big. I think it's till too big right now, but we'll have to take a look together.