Fixed bug in shop where the prefab itself would be modified.
Data is saved and loaded as intended. The prefabs are used to aggregate the default database (i.e. player has no saved data), and clones are created at the shop's initialization which are modified when the item is purchased. The data from these clones is read when saving the player's data.